6axis_revision/requirements.txt
argparse
matplotlib
numpy
pandas
pillow
tensorflow
tqdm
pytorch-lightning
sklearn